==> notebooks/3.0-pb-building-predictive-model.ipynb <==
import pandas as pd
import numpy as np
import os

#Import Data

#set path of processed data
processed_data_path = os.path.join(os.path.pardir,'data','processed')
train_file_path = os.path.join(processed_data_path,'train.csv')
test_file_path = os.path.join(processed_data_path,'test.csv')
    
#Get the data frames
train_df = pd.read_csv(train_file_path, index_col='PassengerId')
test_df = pd.read_csv(test_file_path, index_col='PassengerId')

train_df.info()

test_df.info()

# Data preparation
X = train_df.loc[:,'Age':].values.astype('float')
y = train_df['Survived'].ravel() # ravel gives the vector

print(X.shape,y.shape)

#Train test split
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2,random_state=0) # random_state acts like a seed, to get the same output everytime we split
print(X_train.shape,y_train.shape)
print(y_train.shape,y_test.shape)

print('Mean survival in train : {0:.3f}'.format(np.mean(y_train)))
print('Mean survival in test : {0:.3f}'.format(np.mean(y_test)))

# scikit
import sklearn
sklearn.__version__

#Baseline model
#Basically, baseline model returns the output of the most frequent class


#import function
from sklearn.dummy import DummyClassifier

# create model
model_dummy = DummyClassifier(strategy='most_frequent',random_state=0)


#train model
model_dummy.fit(X_train,y_train)

print('Score for baseline model : {0:.2f}'.format(model_dummy.score(X_test,y_test))) #gives accuracy by default

#performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

#accuracy score
print('Accuracy for baseline model : {0:.2f}'.format(accuracy_score(y_test,model_dummy.predict(X_test))))

#Confusion matrix
print('Confusin matrix for baseline model :\n {0}'.format(confusion_matrix(y_test,model_dummy.predict(X_test))))

#precision score
print('Precision for baseline model : {0:.2f}'.format(precision_score(y_test,model_dummy.predict(X_test))))

#recall score
print('Recall for baseline model : {0:.2f}'.format(recall_score(y_test,model_dummy.predict(X_test))))

# Submitting the baseline model to kagel


# converting to matrix
test_X = test_df.values.astype('float')
print(test_X.shape)

# get predictions
predictions = model_dummy.predict(test_X)

df_submission = pd.DataFrame({'PassengerId':test_df.index,'Survived':predictions})

df_submission.head()

submission_data_path = os.path.join(os.path.pardir,'data','external')
submission_file_path = os.path.join(submission_data_path,'01_dummy.csv')

df_submission.to_csv(submission_file_path, index = False)

def get_submission_file(model,filename):
    #converting to the matrix
    test_X = test_df.values.astype('float')
    
    #make predictions
    predictions = model.predict(test_X)
    
    #submission dataframe
    df_submission = pd.DataFrame({'PassengerId':test_df.index,'Survived':predictions})
    
    #submission file
    submission_data_path = os.path.join(os.path.pardir,'data','external')
    submission_file_path = os.path.join(submission_data_path,filename)
    
    #write to file
    df_submission.to_csv(submission_file_path, index = False)

#get submission file
get_submission_file(model_dummy, '01_dummy.csv')

# Logistic Regression model\
from sklearn.linear_model import LogisticRegression

# create model
model_lr_1 = LogisticRegression(random_state=0)

#train model
model_lr_1.fit(X_train,y_train)

#evaluate model
print('Score for Logistic regression version 1 model : {0:.2f}'.format(model_lr_1.score(X_test,y_test))) #gives accuracy by default

#model coefficients
model_lr_1.coef_

# second submission
get_submission_file(model_lr_1,'02_lr.csv')

# Hyperparameter optimization

#base model
model_lr = LogisticRegression(random_state=0)

from sklearn.model_selection import GridSearchCV


parameters = {'C':[1.0,10.0,50.0,100.0,1000.0],'penalty':['l1','l2']}

clf = GridSearchCV(model_lr,param_grid=parameters,cv=3) # 3-fold cross vaildaiton

clf.fit(X_train,y_train)

clf.best_params_

print('Best score : {0:.2f}'.format(clf.best_score_))

#evaluate model
print('Score for logistic regression version 2 : {0:.2f}'.format(clf.score(X_test,y_test)))

# third submission
get_submission_file(clf,'03_lr.csv')

#Feature normalization and standardiztion

from sklearn.preprocessing import MinMaxScaler, StandardScaler

#Feature normalization
scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_train_scaled[:,0].min(),X_train_scaled[:,0].max()

X_test_scaled = scaler.fit_transform(X_test)

#Feature Standardization
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.fit_transform(X_test)

# Create model after standardization

#base model
model_lr = LogisticRegression()
parameters = {'C':[1.0,10.0,50.0,100.0,1000.0],'penalty':['l1','l2']}
clf = GridSearchCV(model_lr,param_grid = parameters,cv=3)
clf.fit(X_train_scaled,y_train)

clf.best_score_

print('Score for logistic regression after standardization : {0:.2f}'.format(clf.score(X_test_scaled,y_test)))

#model persistence

import pickle

#create the file paths
model_file_path = os.path.join(os.path.pardir,'models','lr_model.pkl')
scaler_file_path = os.path.join(os.path.pardir,'models','lr_scaler.pkl')

#open files to write
model_file_pickle = open(model_file_path,'wb')
scaler_file_pickle = open(scaler_file_path,'wb')

#persist the model and scalar
pickle.dump(clf,model_file_pickle)
pickle.dump(scaler,scaler_file_pickle)

#close the file
model_file_pickle.close()
scaler_file_pickle.close()

# open in read mode
model_file_pickle = open(model_file_path,'rb')
scaler_file_pickle = open(scaler_file_path,'rb')

#Load the persisted files
clf_loaded=pickle.load(model_file_pickle)
scaler_loaded=pickle.load(scaler_file_pickle)

#close the files
model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded


scaler_loaded

X_test_scaled = scaler_loaded.transform(X_test)
print('Score for perssisted logistic regression after standardization : {0:.2f}'.format(clf.score(X_test_scaled,y_test)))

from xgboost import XGBRegressor

my_model = XGBRegressor(n_estimators=1000, learning_rate=0.05)
my_model.fit(X_train, y_train, early_stopping_rounds=5, 
             eval_set=[(X_test, y_test)], verbose=False)


# xgboost submission
get_submission_file(my_model,'04_xgboost.csv')
